Minor cleanup. Split test code.
