+++ .gitignore
@@ -5,4 +5,3 @@ __pycache__/
 build/
 dist/
 .coverage
-.idea/
